refactor particle system
allocating and rendering in this class
